made spike_relevance_analysis more efficient by building
I just build the graph in advance and then I just need to run the
session. I modified sim_pop_activity.py so it just save the packet
for the figure when it should.

+++ model_conv.py
@@ -400,7 +400,7 @@ class WGAN_conv(object):
   #get units STA from the discriminator
   def get_critics_output(self, samples):
       #aux = np.load(self.sample_dir+ '/stats_real.npz')
-      output,_,_ = self.Discriminator_sampler(samples.astype('float32'))
+      output,_,_ = self.Discriminator_sampler(samples)
       return output
    
   #this is to save the network parameters
